add sockets

+++ src/face_detector.py
@@ -20,6 +20,6 @@ def detect_faces(frame):
 
    # detect frontal faces
    face_rects = detect_cascade(frame, FACE_CASCADE, cv2.CASCADE_DO_CANNY_PRUNING)
-   face_points = [[(x1+x2)/2.,(y1+y2)/2.,(np.abs(y2-y1))] for x1, y1, x2, y2 in face_rects]
+   face_points = [[float((x1+x2)/2.),float((y1+y2)/2.),float((np.abs(y2-y1)))] for x1, y1, x2, y2 in face_rects]
 
    return face_rects, face_points
